【优化】 代码优化

+++ app/src/main/java/io/goldstone/blockchain/module/home/rammarket/module/ramtrade/model/TradingInfoModel.kt
@@ -27,7 +27,7 @@ class TradingInfoModel(
 	
 	constructor() : this(
 		"",
-		0.toDouble(),
+		0.0,
 		0,
 		0,
 		0.0,
